fix a race when StopAtEOF is called
Currently when StopAtEOF is called, and we previously encountered an
EOF already, we stop reading the file immediately.  However when
tailing a file, new data might have become available in the meantime,
before the StopAtEOF is called.  The watcher might however not have
notified us about that yet.

Instead of exiting immediately if that happens, and leaving the data
that's already in the file unread, continue iterating until we get
the *next* EOF, as we can be reasonably sure that that's the EOF the
user meant to stop at, making sure to read all the data that has been
written by the time StopAtEOF is called.